Replace map with unordered_map
[Why]
The map does sort the entries, which leads to broken .pc files, since
the order is important, especially for path entries.

[How]
Change the type, unfortunately unordered_map does not provide reverse
iterators, so we have to copy and reverse the order manually.

[Note]
This is also not a 100% guarantee that things will work, since this
depends on the map implementation.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>